python/requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.3.9
cffi==1.15.0
cryptography==37.0.2
PyJWT==2.4.0
requests==2.21.0
urllib3==1.24.2
chardet==3.0.4
future==0.17.1
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
pycparser==2.19
pylint==2.3.1
pyngrok==1.3.7
PyYAML==5.1
six==1.12.0
typed-ast==1.5.4
Werkzeug==0.15.4
wrapt==1.11.1
